Delayed Input
Modified the scripts to introduce an artifical delay between inputs for games that can't accept all input at once. In the future I'd like to add some config to optionally use the delay or not. While keys are being input to the game, control is temporarily disabled for the user - I should look into a more fail-safe solution for this but it works for now.