Case-Insensitive Keys
Minor update to the GameCommands script that maps the command_list.json keys to a lowercase map. This allows both the file and the user to disregard casing and the commands will still work. Also made a minor update to the cooldown duration to prevent awkward cases of durations below 1 second. README has been updated accordingly as well as the example command_list.json file to demonstrate the relaxed casing rules :christmas_tree: